All of it.
We're back baby

+++ code/game/supplyshuttle.dm
@@ -156,7 +156,7 @@ var/list/point_source_descriptions = list(
 	var/list/master_supply_list = list()
 	//shuttle movement
 	var/movetime = 1200
-	var/datum/shuttle/autodock/ferry/supply/shuttle
+	var/datum/shuttle/ferry/supply/cargo/shuttle
 
 	New()
 		ordernum = rand(1,9000)
@@ -193,35 +193,37 @@ var/list/point_source_descriptions = list(
 	proc/sell()
 		var/phoron_count = 0
 		var/plat_count = 0
-		for(var/area/subarea in shuttle.shuttle_area)
-			for(var/atom/movable/MA in subarea)
-				if(MA.anchored)	continue
+		var/area/area_shuttle = shuttle.get_location_area()
+		if(!area_shuttle)	return
 
-				// Must be in a crate!
-				if(istype(MA,/obj/structure/closet/crate))
-					var/obj/structure/closet/crate/CR = MA
-					callHook("sell_crate", list(CR, subarea))
+		for(var/atom/movable/MA in area_shuttle)
+			if(MA.anchored)	continue
 
-					add_points_from_source(CR.points_per_crate, "crate")
-					var/find_slip = 1
+			// Must be in a crate!
+			if(istype(MA,/obj/structure/closet/crate))
+				var/obj/structure/closet/crate/CR = MA
+				callHook("sell_crate", list(CR, area_shuttle))
 
-					for(var/atom in CR)
-						// Sell manifests
-						var/atom/A = atom
-						if(find_slip && istype(A,/obj/item/weapon/paper/manifest))
-							var/obj/item/weapon/paper/manifest/slip = A
-							if(!slip.is_copy && slip.stamped && slip.stamped.len) //yes, the clown stamp will work. clown is the highest authority on the station, it makes sense
-								add_points_from_source(points_per_slip, "manifest")
-								find_slip = 0
-							continue
+				add_points_from_source(CR.points_per_crate, "crate")
+				var/find_slip = 1
 
-						// Sell phoron and platinum
-						if(isstack(A))
-							var/obj/item/stack/P = A
-							switch(P.get_material_name())
-								if("phoron") phoron_count += P.get_amount()
-								if("platinum") plat_count += P.get_amount()
-				qdel(MA)
+				for(var/atom in CR)
+					// Sell manifests
+					var/atom/A = atom
+					if(find_slip && istype(A,/obj/item/weapon/paper/manifest))
+						var/obj/item/weapon/paper/manifest/slip = A
+						if(!slip.is_copy && slip.stamped && slip.stamped.len) //yes, the clown stamp will work. clown is the highest authority on the station, it makes sense
+							add_points_from_source(points_per_slip, "manifest")
+							find_slip = 0
+						continue
+
+					// Sell phoron and platinum
+					if(isstack(A))
+						var/obj/item/stack/P = A
+						switch(P.get_material_name())
+							if("phoron") phoron_count += P.get_amount()
+							if("platinum") plat_count += P.get_amount()
+			qdel(MA)
 
 		if(phoron_count)
 			var/temp = phoron_count * points_per_phoron
@@ -235,17 +237,18 @@ var/list/point_source_descriptions = list(
 	proc/buy()
 		if(!shoppinglist.len) return
 		var/list/clear_turfs = list()
-		for(var/area/subarea in shuttle.shuttle_area)
-			for(var/turf/T in subarea)
-				if(T.density)	continue
-				var/contcount
-				for(var/atom/A in T.contents)
-					if(!A.simulated)
-						continue
-					contcount++
-				if(contcount)
+		var/area/area_shuttle = shuttle.get_location_area()
+		if(!area_shuttle)	return
+		for(var/turf/T in area_shuttle)
+			if(T.density)	continue
+			var/contcount
+			for(var/atom/A in T.contents)
+				if(!A.simulated)
 					continue
-				clear_turfs += T
+				contcount++
+			if(contcount)
+				continue
+			clear_turfs += T
 		for(var/S in shoppinglist)
 			if(!clear_turfs.len)	break
 			var/i = rand(1,clear_turfs.len)
